requirements.txt
Babel==2.9.1
# backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.4.6
botocore==1.6.8
cairocffi==1.1.0
CairoSVG==2.5.1
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.28
django-cors-headers==3.11.0
django-crispy-forms==1.8.0
django-filter==21.1
# django-heroku==0.3.1
django-paypal==1.0.0
django-phonenumber-field==3.0.1
django-sslserver==0.22
django-storages==1.8
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.15.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
Js2Py==0.74
packaging==23.0
phonenumbers==8.10.21
Pillow==9.0.1
pipwin==0.5.2
# psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.19
pyjsparser==2.7.1
Pyphen==0.9.5
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
requests==2.22.0
s3transfer==0.1.13
six==1.13.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
tinycss2==1.0.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.25.5
uuid==1.30
WeasyPrint==50
webencodings==0.5.1
whitenoise==5.0.1
